chore: organize table component
