+++ src/pages/marketplaces/index.tsx
@@ -1,7 +1,7 @@
 import { Accounts as Icon } from '@/assets/title-icons';
 import Copy from '@/components/Copy';
 import Title from '@/components/Layout/Title';
-import Table, { ITable } from '@/components/TableV2';
+import Table, { ITable } from '@/components/Table';
 import { getMarketplaces } from '@/services/requests/marketplace';
 import { Container, FlexSpan, Header, Mono } from '@/styles/common';
 import { IRowSection } from '@/types';
