+++ src/pages/validators/index.tsx
@@ -4,8 +4,8 @@ import Copy from '@/components/Copy';
 import Detail from '@/components/Detail';
 import { IFilter } from '@/components/Filter';
 import Progress from '@/components/Progress';
-import { ITable } from '@/components/TableV2';
-import { CustomFieldWrapper, Status } from '@/components/TableV2/styles';
+import { ITable } from '@/components/Table';
+import { CustomFieldWrapper, Status } from '@/components/Table/styles';
 import Tooltip from '@/components/Tooltip';
 import api from '@/services/api';
 import { CenteredRow, DoubleRow, Mono } from '@/styles/common';
