+++ src/views/assets/index.ts
@@ -1,5 +1,3 @@
-import { ITableType } from '@/components/Table/styles';
-import widths from '@/components/Table/widths';
 import { CardContent, CardHeaderItem, Container } from '@/styles/common';
 import styled, { css } from 'styled-components';
 
@@ -17,71 +15,6 @@ export const Title = styled.div`
     svg {
       height: auto;
       width: auto;
-    }
-  }
-`;
-
-export const Row = styled.div<ITableType>`
-  padding: 1rem 1.5rem;
-
-  display: flex;
-
-  flex-direction: row;
-  align-items: center;
-
-  background-color: ${props => props.theme.white};
-
-  border-radius: 0.5rem;
-
-  div {
-    margin-right: 0.3125rem;
-  }
-
-  span {
-    &:nth-child(1) {
-      margin: -10px 33px -10px -10px;
-    }
-  }
-
-  span,
-  a {
-    /* flex: 1; */
-    overflow: hidden;
-
-    text-overflow: ellipsis;
-    white-space: nowrap;
-
-    font-size: 0.95rem;
-    color: ${props => props.theme.black};
-
-    ${props => widths[props.type]};
-
-    a {
-      color: ${props => props.theme.black};
-      font-weight: 600;
-    }
-
-    small {
-      color: ${props => props.theme.darkText};
-    }
-
-    strong {
-      font-weight: 400;
-      font-size: 0.95rem;
-      color: ${props => props.theme.darkText};
-    }
-
-    p {
-      font-weight: 600;
-      color: ${props => props.theme.black};
-    }
-  }
-  .address {
-    cursor: pointer;
-    text-decoration: none;
-    font-weight: 500;
-    &:hover {
-      text-decoration: underline;
     }
   }
 `;
